ugc_api/requirements.txt
fastapi
uvicorn[standard]
pydantic<2
python-jose[cryptography]
aiokafka
python-logstash
requests